+++ AdapterUnitTests/MintegralAdapterTests/AUTMintegralRTBRewardedAdTests.m
@@ -136,8 +136,6 @@ static NSString *const kBidResponse = @"bidResponse";
       .andDo(^(NSInvocation *invocation) {
         [self->_adLoader onVideoAdShowSuccess:kPlacementID unitId:kUnitID];
       });
-  OCMStub([_rewardedAdMock isVideoReadyToPlayWithPlacementId:kPlacementID unitId:kUnitID])
-      .andReturn(YES);
   AUTKMediationRewardedAdEventDelegate *eventDelegate = [self loadAd];
   // Assert the initial values of the counts before they are verified after the "Act" steps.
   XCTAssertEqual(eventDelegate.willPresentFullScreenViewInvokeCount, 0);
@@ -169,8 +167,6 @@ static NSString *const kBidResponse = @"bidResponse";
       .andDo(^(NSInvocation *invocation) {
         [self->_adLoader onVideoAdShowSuccess:kPlacementID unitId:kUnitID];
       });
-  OCMStub([_rewardedAdMock isVideoReadyToPlayWithPlacementId:kPlacementID unitId:kUnitID])
-      .andReturn(YES);
   AUTKMediationRewardedAdEventDelegate *eventDelegate = [self loadAd];
   // Assert the initial values of the counts before they are verified after the "Act" steps.
   XCTAssertEqual(eventDelegate.willPresentFullScreenViewInvokeCount, 0);
@@ -191,16 +187,6 @@ static NSString *const kBidResponse = @"bidResponse";
   XCTAssertEqual(eventDelegate.didDismissFullScreenViewInvokeCount, 1);
 }
 
-- (void)testShowFailureForAdNotReadyToShow {
-  OCMStub([_rewardedAdMock isVideoReadyToPlayWithPlacementId:kPlacementID unitId:kUnitID])
-      .andReturn(NO);
-  AUTKMediationRewardedAdEventDelegate *eventDelegate = [self loadAd];
-
-  [_adLoader presentFromViewController:[[UIViewController alloc] init]];
-
-  XCTAssertEqual(eventDelegate.didFailToPresentError.code, GADMintegralErrorAdFailedToShow);
-}
-
 - (void)testShowFailure {
   NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterMintegralErrorDomain
                                                       code:GADMintegralErrorAdFailedToShow
@@ -215,8 +201,6 @@ static NSString *const kBidResponse = @"bidResponse";
       .andDo(^(NSInvocation *invocation) {
         [self->_adLoader onVideoAdShowFailed:kPlacementID unitId:kUnitID withError:expectedError];
       });
-  OCMStub([_rewardedAdMock isVideoReadyToPlayWithPlacementId:kPlacementID unitId:kUnitID])
-      .andReturn(YES);
   AUTKMediationRewardedAdEventDelegate *eventDelegate = [self loadAd];
   XCTAssertEqual(eventDelegate.willPresentFullScreenViewInvokeCount, 0);