+++ AdapterUnitTests/MintegralAdapterTests/AUTMintegralWaterfallInterstitialAdTests.m
@@ -120,7 +120,6 @@ static NSString *const kUnitID = @"67890";
       .andDo(^(NSInvocation *invocation) {
         [self->_adLoader newInterstitialAdShowSuccess:self->_interstitialAdMock];
       });
-  OCMStub([_interstitialAdMock isAdReady]).andReturn(YES);
   AUTKMediationInterstitialAdEventDelegate *eventDelegate = [self loadAd];
   // Assert the initial values of the counts before they are verified after the "Act" steps.
   XCTAssertEqual(eventDelegate.willPresentFullScreenViewInvokeCount, 0);
@@ -141,15 +140,6 @@ static NSString *const kUnitID = @"67890";
   XCTAssertEqual(eventDelegate.didDismissFullScreenViewInvokeCount, 1);
 }
 
-- (void)testShowFailureForAdNotReadyToShow {
-  OCMStub([_interstitialAdMock isAdReady]).andReturn(NO);
-  AUTKMediationInterstitialAdEventDelegate *eventDelegate = [self loadAd];
-
-  [_adLoader presentFromViewController:[[UIViewController alloc] init]];
-
-  XCTAssertEqual(eventDelegate.didFailToPresentError.code, GADMintegralErrorAdFailedToShow);
-}
-
 - (void)testShowFailureForAdShowFail {
   NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterMintegralErrorDomain
                                                       code:GADMintegralErrorAdFailedToShow
@@ -160,7 +150,6 @@ static NSString *const kUnitID = @"67890";
         [self->_adLoader newInterstitialAdShowFail:expectedError
                                          adManager:self->_interstitialAdMock];
       });
-  OCMStub([_interstitialAdMock isAdReady]).andReturn(YES);
   AUTKMediationInterstitialAdEventDelegate *eventDelegate = [self loadAd];
   XCTAssertEqual(eventDelegate.willPresentFullScreenViewInvokeCount, 0);